添加记住用户勾选项，并使用cookie记住用户

+++ app/controllers/sessions_controller.rb
@@ -5,6 +5,7 @@ class SessionsController < ApplicationController
   	user = User.find_by(email: params[:session][:email].downcase)
   	if user && user.authenticate(params[:session][:password])
   		login(user)
+  		remember(user) if params[:session][:remember_me] == '1'
   		redirect_to user
   	else
   		flash.now[:danger] = '用户名不存在或密码不正确' 
@@ -13,8 +14,13 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-  	session.delete(:user_id)
-  	current_user = nil
-  	redirect_to root_path
+  	if login?
+  		current_user.update_attribute(:remember_digest, nil)
+  		cookies.delete(:user_id)
+  		cookies.delete(:remember_token)
+  		session.delete(:user_id)
+  		current_user = nil
+  		redirect_to root_path
+  	end
   end
 end
